Add update and create functionality for recipe relationships - Implement handler for creating and updating recipes - Add SQL DTO for recipe ingredients (create, update) - Implement MySQL storage for creating and updating recipe ingredients - Implement business logic for recipe creation and updates

+++ internal/entity/dtos/redis/ingredient/create_ingredient.go
@@ -3,7 +3,7 @@ package ingredientcachemodel
 import "tart-shop-manager/internal/common"
 
 type CreateIngredient struct {
-	IngredientID uint   `gorm:"column:ingredient_id;primaryKey;autoIncrement" json:"ingredient_id"`
+	IngredientID uint64 `gorm:"column:ingredient_id;primaryKey;autoIncrement" json:"ingredient_id"`
 	Name         string `gorm:"column:name;size:200;not null" json:"name"`
 	Description  string `gorm:"column:description;type:text" json:"description"`
 	Unit         string `gorm:"column:unit;size:100;not null" json:"unit"`
